Write documentation for Server and cleanup
