--- src/main/kotlin/simulation/exceptions/ServerIsAlreadyBlockedException.kt
@@ -1,9 +1,0 @@
-/*
- * Copyright (c) 2023. Andrés Arango Pérez <arangoandres.p@gmail.com>
- *
- * You may use, distribute and modify this code under the terms of the MIT license.
- */
-
-package simulation.exceptions
-
-class ServerIsAlreadyBlockedException : Throwable()
